# Wrap in/out length pointer APIs: avoid heap escape