* Related: https://github.com/golang-fips/openssl/pull/129
* For: https://github.com/golang-fips/openssl/issues/128

Add wrappers that help avoid escape to the heap for this list:

https://github.com/golang-fips/openssl/blob/61234a910b77a1b4cabccee389684cbc372fd5ea/cgo_go122.go#L10-L17

There's only a benchmark for ECDH (EVP_PKEY_derive) but this PR shows the same results as https://github.com/golang-fips/openssl/pull/113:

```
goos: linux
goarch: amd64
pkg: github.com/golang-fips/openssl/v2
cpu: Intel(R) Core(TM) i7-6700 CPU @ 3.40GHz
       │   old.txt   │           new.txt            │
       │   sec/op    │   sec/op     vs base         │
ECDH-8   104.1µ ± 3%   101.2µ ± 3%  ~ (p=0.065 n=8)

       │  old.txt   │              new.txt              │
       │    B/op    │    B/op     vs base               │
ECDH-8   40.00 ± 0%   32.00 ± 0%  -20.00% (p=0.000 n=8)

       │  old.txt   │              new.txt              │
       │ allocs/op  │ allocs/op   vs base               │
ECDH-8   2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=8)
```

---

The wrapper makes it so you can't distinguish a "valid" returned `0` vs. a `0` returned because there's an error, and I don't know for sure that won't cause problems in some use cases. (E.g. make an error show up in the wrong way.)

Returning a C struct with two values is what I have in mind make a more transparent wrapper that avoids losing this difference, but I figured I'd submit this PR first in case it's not worth the complexity. Please let me know if that seems better. (The benchmark result is the same when I tried this out for `EVP_PKEY_derive`.)